Add LogTestSymbolProvider for log file symbol navigation
- Implemented a DocumentSymbolProvider for 'log' files to extract and provide symbols for test names and test sequences.
- Added regex patterns to match lines indicating the start of tests and test sequences.
- Registered the symbol provider in the extension activation function.
